fix(backend): modelToCRDTrigger was not including periodic schedule correctly
