+++ backend/src/apiserver/model/periodic_schedule.go
@@ -1,0 +1,21 @@
+package model
+
+type PeriodicSchedule struct {
+	// Time at which scheduling starts.
+	// If no start time is specified, the StartTime is the creation time of the schedule.
+	PeriodicScheduleStartTimeInSec *int64 `gorm:"column:PeriodicScheduleStartTimeInSec;"`
+
+	// Time at which scheduling ends.
+	// If no end time is specified, the EndTime is the end of time.
+	PeriodicScheduleEndTimeInSec *int64 `gorm:"column:PeriodicScheduleEndTimeInSec;"`
+
+	// Interval describing when a workflow should be created within the
+	// time interval defined by StartTime and EndTime.
+	IntervalSecond *int64 `gorm:"column:IntervalSecond;"`
+}
+
+func (p PeriodicSchedule) IsEmpty() bool {
+	return p.PeriodicScheduleStartTimeInSec == nil &&
+		p.PeriodicScheduleEndTimeInSec == nil &&
+		(p.IntervalSecond == nil || *p.IntervalSecond == 0)
+}
